feat: refactor oauth screen
